Cập nhật các chức năng khác

+++ lib/presentation/widgets/separator.dart
@@ -1,22 +1,21 @@
-import 'package:flutter/cupertino.dart';
-import 'package:movie_app/common/constants/size_constants.dart';
-import 'package:movie_app/common/extensions/size_extensions.dart';
-import 'package:movie_app/presentation/themes/theme_color.dart';
+import 'package:flutter/material.dart';
+
+import '../../common/constants/size_constants.dart';
+import '../../common/extensions/size_extensions.dart';
+import '../themes/theme_color.dart';
 
 class Separator extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Container(
-      height: Sizes.dimen_1.h.toDouble(),  // 🔥 Thêm .toDouble()
-      width: Sizes.dimen_80.w.toDouble(), // 🔥 Thêm .toDouble()
+      height: Sizes.dimen_1.h,
+      width: Sizes.dimen_80.w,
       padding: EdgeInsets.only(
-        top: Sizes.dimen_2.h.toDouble(),  // 🔥 Thêm .toDouble()
-        bottom: Sizes.dimen_6.h.toDouble(), // 🔥 Thêm .toDouble()
+        top: Sizes.dimen_2.h,
+        bottom: Sizes.dimen_6.h,
       ),
       decoration: BoxDecoration(
-        borderRadius: BorderRadius.all(
-          Radius.circular(Sizes.dimen_1.h.toDouble()), // 🔥 Thêm .toDouble()
-        ),
+        borderRadius: BorderRadius.all(Radius.circular(Sizes.dimen_1.h)),
         gradient: LinearGradient(
           colors: [
             AppColor.violet,
